Add files via upload
I will present to you a project I developed, which is a management system for donations and humanitarian contributions to orphans. This project is an online interface where users can introduce themselves, choose to donate to a specific or general orphan, and choose the types of assistance they would like to offer. The program includes monetary, financial and educational donation options and is designed to provide simple and seamless user interaction.